# swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The LocalAI Rest API.",
        "title": "LocalAI API",
        "contact": {
            "name": "LocalAI",
            "url": "https://localai.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://raw.githubusercontent.com/mudler/LocalAI/master/LICENSE"
        },
        "version": "2.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/instructions/{name}": {
            "get": {
                "description": "Returns a markdown guide (default) or filtered OpenAPI fragment (format=json) for a named instruction",
                "produces": [
                    "application/json",
                    "text/markdown"
                ],
                "tags": [
                    "instructions"
                ],
                "summary": "Get an instruction's API guide or OpenAPI fragment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instruction name (e.g. chat-inference, config-management)",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Response format: json for OpenAPI fragment, omit for markdown",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "instruction documentation",
                        "schema": {
                            "$ref": "#/definitions/localai.APIInstructionResponse"
                        }
                    },
                    "404": {
                        "description": "instruction not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "localai.APIInstructionResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "swagger_fragment": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}